+++ eminus/orbitals.py
@@ -4,8 +4,9 @@
 
 from .dft import get_psi
 from .io import write_cube
-from .localizer import get_FLO, get_FO, get_wannier
+from .localizer import get_FLO, get_FO, get_scdm, get_wannier
 from .logger import log
+from .tools import orbital_center
 
 
 def KSO(scf, write_cubes=False, **kwargs):
@@ -32,7 +33,7 @@ def KSO(scf, write_cubes=False, **kwargs):
     return kso
 
 
-def FO(scf, write_cubes=False, fods=None):
+def FO(scf, write_cubes=False, fods=None, guess='wannier'):
     """Generate Fermi orbitals and optionally save them as CUBE files.
 
     Reference: J. Chem. Phys. 153, 084104.
@@ -43,6 +44,7 @@ def FO(scf, write_cubes=False, fods=None):
     Keyword Args:
         write_cubes: Write orbitals to CUBE files.
         fods: Fermi-orbital descriptors.
+        guess: Guess to generate FODs if none are given. Can be 'Wannier' (default) or 'PyCOM'.
 
     Returns:
         Real-space Fermi orbitals.
@@ -54,9 +56,12 @@ def FO(scf, write_cubes=False, fods=None):
 
     # Calculate eigenfunctions
     kso = get_psi(scf, scf.W)
-    if fods is None:
+    if fods is None and guess == 'wannier':
+        wo = WO(scf)
+        fods = orbital_center(atoms, wo[0])
+    if fods is None and guess == 'pycom':
         fods = get_fods(atoms)
-    fods = remove_core_fods(atoms, fods)
+        fods = remove_core_fods(atoms, fods)
 
     # The FO functions need orbitals in reciprocal space as input
     fo = get_FO(atoms, kso, fods)
@@ -65,7 +70,7 @@ def FO(scf, write_cubes=False, fods=None):
     return fo
 
 
-def FLO(scf, write_cubes=False, fods=None):
+def FLO(scf, write_cubes=False, fods=None, guess='wannier'):
     """Generate Fermi-Loewdin orbitals and optionally save them as CUBE files.
 
     Reference: J. Chem. Phys. 153, 084104.
@@ -76,6 +81,7 @@ def FLO(scf, write_cubes=False, fods=None):
     Keyword Args:
         write_cubes: Write orbitals to CUBE files.
         fods: Fermi-orbital descriptors.
+        guess: Guess to generate FODs if none are given. Can be 'Wannier' (default) or 'PyCOM'.
 
     Returns:
         Real-space Fermi-Loewdin orbitals.
@@ -87,9 +93,14 @@ def FLO(scf, write_cubes=False, fods=None):
 
     # Calculate eigenfunctions
     kso = get_psi(scf, scf.W)
-    if fods is None:
+
+    guess = guess.lower()
+    if fods is None and guess == 'wannier':
+        wo = WO(scf)
+        fods = orbital_center(atoms, wo[0])
+    if fods is None and guess == 'pycom':
         fods = get_fods(atoms)
-    fods = remove_core_fods(atoms, fods)
+        fods = remove_core_fods(atoms, fods)
 
     # The FLO functions need orbitals in reciprocal space as input
     flo = get_FLO(atoms, kso, fods)
@@ -108,7 +119,7 @@ def WO(scf, write_cubes=False, precondition=True):
 
     Keyword Args:
         write_cubes: Write orbitals to CUBE files.
-        precondition: Precondition by calculating FLOs as the initial guess.
+        precondition: Precondition by calculating SCDMOs as the initial guess.
 
     Returns:
         Real-space Wannier orbitals.
@@ -117,7 +128,7 @@ def WO(scf, write_cubes=False, precondition=True):
 
     # Calculate eigenfunctions/initial guess orbitals and transform to real-space
     if precondition:
-        psi = FLO(scf)
+        psi = SCDMO(scf)
     else:
         psi = atoms.I(get_psi(scf, scf.W))
 
@@ -125,6 +136,30 @@ def WO(scf, write_cubes=False, precondition=True):
     if write_cubes:
         cube_writer(atoms, 'WO', wo)
     return wo
+
+
+def SCDMO(scf, write_cubes=False):
+    """Generate SCDM orbitals and optionally save them as CUBE files.
+
+    Reference: J. Comput. Phys. 334, 1.
+
+    Args:
+        scf: SCF object.
+
+    Keyword Args:
+        write_cubes: Write orbitals to CUBE files.
+
+    Returns:
+         Real-space SCDM orbitals.
+    """
+    atoms = scf.atoms
+
+    # Calculate eigenfunctions
+    kso = get_psi(scf, scf.W)
+    scdmo = get_scdm(atoms, kso)
+    if write_cubes:
+        cube_writer(atoms, 'SCDMO', scdmo)
+    return scdmo
 
 
 def cube_writer(atoms, orb_type, orbitals):