+++ tests/test_io.py
@@ -207,7 +207,6 @@ def test_write_method():
 
 if __name__ == '__main__':
     import inspect
-    import pathlib
 
     file_path = pathlib.Path(inspect.stack()[0][1])
     pytest.main(file_path)